Code tidy

+++ src/main/java/org/ek9lang/core/utils/OsSupport.java
@@ -3,25 +3,21 @@ package org.ek9lang.core.utils;
 import org.ek9lang.core.exception.AssertValue;
 
 import java.io.File;
-import java.io.FilenameFilter;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 
 public class OsSupport
 {
 	public static int numberOfProcessors()
 	{
-		int processors = Runtime.getRuntime().availableProcessors();
-	    return processors;
+		return Runtime.getRuntime().availableProcessors();
 	}
 
 	public String getUsersHomeDirectory()
 	{
 		return System.getProperty("user.home");
 	}
-	
+
 	public String getCurrentWorkingDirectory()
 	{
 		return System.getProperty("user.dir");
@@ -31,12 +27,12 @@ public class OsSupport
 	{
 		return OsSupport.numberOfProcessors();
 	}
-	
+
 	public String getFileNameWithoutPath(String fileNameWithPath)
 	{
 		if(fileNameWithPath == null || fileNameWithPath.isEmpty())
 			return "";
-		
+
 		File f = new File(fileNameWithPath);
 		return f.getName();
 	}
@@ -45,35 +41,36 @@ public class OsSupport
 	{
 		return fileName != null && isFileReadable(new File(fileName));
 	}
-	
+
 	public boolean isFileReadable(File file)
-	{		
+	{
 		return file != null && file.isFile() && !file.isDirectory() && file.canRead();
 	}
-	
+
 	public boolean isDirectoryReadable(String directoryName)
 	{
 		return directoryName != null && isDirectoryReadable(new File(directoryName));
 	}
-	
+
 	public boolean isDirectoryReadable(File directory)
 	{
 		return directory != null && directory.isDirectory() && directory.canRead();
 	}
-	
+
 	public boolean isDirectoryWritable(String directoryName)
 	{
 		return isDirectoryWritable(new File(directoryName));
 	}
-	
+
 	public boolean isDirectoryWritable(File directory)
 	{
 		return directory != null && directory.isDirectory() && directory.canWrite();
 	}
-	
+
 	/**
 	 * deletes matching files
-	 * @param dir the directory to look in
+	 *
+	 * @param dir             the directory to look in
 	 * @param fileNamePattern The pattern regex not shell so for * use .* for .txt use \\.txt
 	 */
 	public void deleteMatchingFiles(File dir, String fileNamePattern)
@@ -81,56 +78,58 @@ public class OsSupport
 		AssertValue.checkNotNull("Dir cannot be null", dir);
 		AssertValue.checkNotNull("FileNamePattern cannot be null", fileNamePattern);
 
-		File[] files = dir.listFiles( new FilenameFilter() {
-			@Override
-			public boolean accept( final File dir, final String name )
+		File[] files = dir.listFiles((dir1, name) -> name.matches(fileNamePattern));
+		if(files != null)
+		{
+			for(final File file : files)
 			{
-				return name.matches(fileNamePattern);
+				if(!file.delete())
+					System.err.println("Can't remove " + file.getAbsolutePath());
 			}
-		});
-		for(final File file : files)
-			if (!file.delete())
-			{
-				System.err.println("Can't remove " + file.getAbsolutePath());
-			}
+		}
 	}
+
 	/**
 	 * Does a recursive delete from this directory and below.
 	 * If includeDirectoryRoot is true then it will delete that directory as well
 	 */
+	@SuppressWarnings("ResultOfMethodCallIgnored")
 	public void deleteContentsAndBelow(File dir, boolean includeDirectoryRoot)
 	{
-
-		for(File toDelete : dir.listFiles())
-		{			
-			if(toDelete.isDirectory())
-				deleteContentsAndBelow(toDelete, true);
-			else
-				toDelete.delete();
+		File[] files = dir.listFiles();
+		if(files != null)
+		{
+			for(File toDelete : files)
+			{
+				if(toDelete.isDirectory())
+					deleteContentsAndBelow(toDelete, true);
+				else
+					toDelete.delete();
+			}
 		}
 		if(includeDirectoryRoot)
 			dir.delete();
 	}
-	
+
 	public List<File> getFilesFromDirectories(Collection<File> inDirectories, String fileSuffix)
 	{
 		AssertValue.checkNotNull("InDirectories cannot be null", inDirectories);
 		AssertValue.checkNotNull("FileSuffix cannot be null", fileSuffix);
 
-		ArrayList<File> rtn = new ArrayList<File>();
+		ArrayList<File> rtn = new ArrayList<>();
 		inDirectories.forEach(dir -> rtn.addAll(getFilesFromDirectory(dir, fileSuffix)));
 		return rtn;
 	}
-	
+
 	public List<File> getDirectoriesInDirectory(File inDirectory, String excludeStartingWith)
 	{
 		AssertValue.checkNotNull("InDirectory cannot be null", inDirectory);
 		AssertValue.checkNotNull("ExcludeStartingWith cannot be null", excludeStartingWith);
 
-		ArrayList<File> rtn = new ArrayList<File>();
+		ArrayList<File> rtn = new ArrayList<>();
 		File[] files = inDirectory.listFiles((d, name) -> !name.startsWith(excludeStartingWith));
 		if(files != null)
-			for (File f : files)
+			for(File f : files)
 				if(f.isDirectory())
 					rtn.add(f);
 		return rtn;
@@ -141,7 +140,7 @@ public class OsSupport
 		AssertValue.checkNotNull("InDirectory cannot be null", inDirectory);
 		AssertValue.checkNotNull("SearchCondition cannot be null", searchCondition);
 
-		ArrayList<File> rtn = new ArrayList<File>();
+		ArrayList<File> rtn = new ArrayList<>();
 		List<File> fullList = getFilesRecursivelyFrom(inDirectory);
 		fullList.forEach(file -> {
 			Path thePath = file.toPath();
@@ -151,16 +150,16 @@ public class OsSupport
 		});
 		return rtn;
 	}
-	
+
 	public List<File> getFilesRecursivelyFrom(File inDirectory)
 	{
 		AssertValue.checkNotNull("InDirectory cannot be null", inDirectory);
 
-		ArrayList<File> rtn = new ArrayList<File>();
+		ArrayList<File> rtn = new ArrayList<>();
 		File[] files = inDirectory.listFiles();
 		if(files != null)
 		{
-			for (File f : files)
+			for(File f : files)
 			{
 				if(f.isDirectory())
 					rtn.addAll(getFilesRecursivelyFrom(f));
@@ -170,52 +169,48 @@ public class OsSupport
 		}
 		return rtn;
 	}
-	
+
 	public Collection<File> getFilesFromDirectory(File inDirectory, String fileSuffix)
 	{
 		AssertValue.checkNotNull("InDirectory cannot be null", inDirectory);
 		AssertValue.checkNotNull("FileSuffix cannot be null", fileSuffix);
 
-		ArrayList<File> rtn = new ArrayList<File>();
+		ArrayList<File> rtn = new ArrayList<>();
 		File[] files = inDirectory.listFiles((d, name) -> name.endsWith(fileSuffix));
 		if(files != null)
-			for(File f: files)
-				rtn.add(f);
-		
+			Collections.addAll(rtn, files);
+
 		return rtn;
 	}
-	
+
 	public Collection<File> getAllSubdirectories(String directoryRoot)
 	{
 		AssertValue.checkNotNull("DirectoryRoot cannot be null", directoryRoot);
 
-		ArrayList<File> rtn = new ArrayList<File>();
-		
+		ArrayList<File> rtn = new ArrayList<>();
+
 		File dir = new File(directoryRoot);
 		rtn.add(dir);
 		rtn.addAll(doGetAllSubdirectories(dir));
-		
+
 		return rtn;
 	}
-	
+
 	private Collection<File> doGetAllSubdirectories(File dir)
 	{
 		AssertValue.checkNotNull("Dir cannot be null", dir);
 
-		ArrayList<File> rtn = new ArrayList<File>();
-		
+		ArrayList<File> rtn = new ArrayList<>();
+
 		File[] files = dir.listFiles();
-		
 		if(files != null)
 		{
-			for(File f: files)
-			{
-				if(f.isDirectory() && f.canRead())
-				{
-					rtn.add(f);
-					rtn.addAll(doGetAllSubdirectories(f));
-				}
-			}
+			Arrays.stream(files)
+					.filter(file -> file.isDirectory() && file.canRead())
+					.forEach(file -> {
+						rtn.add(file);
+						rtn.addAll(doGetAllSubdirectories(file));
+					});
 		}
 		return rtn;
 	}
